chore: 스토리북 preview에 global styles, decorators  설정 추가 (#13)
