+++ package.json
@@ -48,6 +48,7 @@
     "@storybook/addon-docs": "^9.1.1",
     "@storybook/addon-onboarding": "^9.1.1",
     "@storybook/addon-vitest": "^9.1.1",
+    "@storybook/react": "^9.1.2",
     "@storybook/react-vite": "^9.1.1",
     "@svgr/webpack": "^8.1.0",
     "@types/react": "^19.1.8",
